+++ app/src/main/java/org/mozilla/fenix/HomeActivity.kt
@@ -27,6 +27,7 @@ import org.mozilla.fenix.components.metrics.Event
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.home.HomeFragmentDirections
 import org.mozilla.fenix.library.bookmarks.BookmarkFragmentDirections
+import org.mozilla.fenix.library.bookmarks.selectfolder.SelectBookmarkFolderFragmentDirections
 import org.mozilla.fenix.library.history.HistoryFragmentDirections
 import org.mozilla.fenix.search.SearchFragmentDirections
 import org.mozilla.fenix.settings.SettingsFragmentDirections
@@ -155,6 +156,8 @@ open class HomeActivity : AppCompatActivity() {
                 SettingsFragmentDirections.actionSettingsFragmentToBrowserFragment(sessionId)
             BrowserDirection.FromBookmarks ->
                 BookmarkFragmentDirections.actionBookmarkFragmentToBrowserFragment(sessionId)
+            BrowserDirection.FromBookmarksFolderSelect ->
+                SelectBookmarkFolderFragmentDirections.actionBookmarkSelectFolderFragmentToBrowserFragment(sessionId)
             BrowserDirection.FromHistory ->
                 HistoryFragmentDirections.actionHistoryFragmentToBrowserFragment(sessionId)
         }
@@ -193,5 +196,5 @@ open class HomeActivity : AppCompatActivity() {
 }
 
 enum class BrowserDirection {
-    FromGlobal, FromHome, FromSearch, FromSettings, FromBookmarks, FromHistory
+    FromGlobal, FromHome, FromSearch, FromSettings, FromBookmarks, FromBookmarksFolderSelect, FromHistory
 }